fe: rename hook to be more generic
update API call to new golang backend
